# docs/architecture/api-specification.yaml
openapi: 3.0.3
info:
  title: FitrahTube API
  version: 1.0.0
  description: |
    REST API for FitrahTube backend. Production implementation with 113+ endpoints across 14 controllers.

    **Key Features:**
    - Public content API for Android app (no auth required)
    - Admin/moderator content management with approval workflows
    - YouTube integration for content discovery
    - Bulk import/export functionality
    - Video validation and health monitoring
    - Comprehensive audit logging

    **Authentication:**
    - Firebase Authentication with JWT tokens
    - Role-based access: ADMIN, MODERATOR
    - Public endpoints require no authentication

    **Pagination:**
    - Cursor-based pagination for infinite scroll
    - Page-based pagination for admin interfaces

    **Caching:**
    - Categories: 1 hour TTL
    - Channels: 15 minutes TTL
    - YouTube searches: Varies by type
servers:
  - url: http://localhost:8080/api
    description: Local Development
  - url: http://192.168.1.167:8080/api
    description: Local Network (Android Testing)
  - url: https://api.albunyaan.com/api
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/content:
    get:
      summary: Get paginated content for mobile app
      description: |
        Fetch content filtered by type (HOME, CHANNELS, PLAYLISTS, VIDEOS).
        Returns cursor-paginated results for infinite scroll.
      operationId: getPublicContent
      tags: [Public Content]
      security: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [HOME, CHANNELS, PLAYLISTS, VIDEOS]
            default: HOME
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: category
          in: query
          description: Filter by category slug
          schema:
            type: string
        - name: length
          in: query
          schema:
            type: string
            enum: [SHORT, MEDIUM, LONG]
        - name: date
          in: query
          schema:
            type: string
            enum: [LAST_24_HOURS, LAST_7_DAYS, LAST_30_DAYS]
        - name: sort
          in: query
          schema:
            type: string
            enum: [DEFAULT, MOST_POPULAR, NEWEST]
      responses:
        '200':
          description: Content page
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CursorPageDto'
                  - type: object
                    description: Cursor-paginated mixed content response
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ContentItemDto'
...
  /v1/home:
    get:
      summary: Get paginated home feed with category sections
      description: |
        Returns categories in admin-defined displayOrder, each containing up to
        contentLimit items in admin-defined sort order. Supports cursor-based
        pagination for infinite scroll of category sections.
      operationId: getHomeFeed
      tags: [Public Content]
      security: []
      parameters:
        - name: cursor
          in: query
          description: Base64-encoded cursor from previous page
          schema:
            type: string
        - name: categoryLimit
          in: query
          description: Max categories per page
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 5
        - name: contentLimit
          in: query
          description: Max items per category
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
      responses:
        '200':
          description: Paginated home feed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CursorPageDto'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HomeCategoryDto'
  /v1/channels/{channelId}:
    get:
      summary: Get channel details by YouTube ID
      operationId: getPublicChannel
      tags: [Public Content]
      security: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          description: YouTube channel ID
      responses:
        '200':
          description: Channel entity (without nested playlists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '404':
          description: Channel not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/playlists/{playlistId}:
    get:
      summary: Get playlist details by YouTube ID
      operationId: getPublicPlaylist
      tags: [Public Content]
      security: []
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: string
          description: YouTube playlist ID
      responses:
        '200':
          description: Playlist entity (without nested videos)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/videos/{videoId}:
    get:
      summary: Get video details by YouTube ID
      operationId: getPublicVideo
      tags: [Public Content]
      security: []
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
          description: YouTube video ID
      responses:
        '200':
          description: Video entity with metadata and approval status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        '404':
          description: Video not found or not approved/available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/search:
    get:
      summary: Search across all content types
      operationId: searchPublicContent
      tags: [Public Content]
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [CHANNELS, PLAYLISTS, VIDEOS]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentItemDto'
                description: Array of matching content items (not paginated)

  # ============================================================================
  # SORT ORDER (Admin)
  # ============================================================================
...
  /admin/sort/categories/{categoryId}/content:
    get:
      summary: Get content items in sort order within a category
      operationId: getCategoryContentOrder
      tags: [Sort Order]
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Content items sorted by position
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentSortDto'
...
  /admin/content:
    get:
      summary: Get all content with filtering
      description: |
        Retrieves content items (channels, playlists, videos) with support for filtering,
        searching, and pagination.

        **Search Behavior:**
        Search matches against title, keywords/tags, and description fields using
        case-insensitive contains matching. This enables finding content by:
        - Partial title text (e.g., "islam" matches "Introduction to Islam")
        - Keyword tags (e.g., "quran" matches items tagged with "quran recitation")
        - Description text (e.g., "arabic" matches items with "arabic" in description)

        **Query Limits (Important):**
        To prevent Firestore quota exhaustion, this endpoint uses bounded queries that
        limit results to a maximum of 200 items per content type. When the `truncated`
        field in the response is `true`, there may be more matching items than shown.

        **Best Practices for Large Datasets:**
        - Use specific type filters (`types=channel`) instead of fetching all types
        - Apply category and status filters to narrow results
        - Use pagination (`page`, `size`) for browsing
        - For finding specific items, use the search parameter with distinctive terms
      operationId: listContent
      tags: [Content Library]
      parameters:
        - name: types
          in: query
          description: Comma-separated content types to include. Defaults to all types.
          schema:
            type: string
            example: channel,playlist,video
        - name: status
          in: query
          description: Filter by approval status
          schema:
            type: string
            enum: [all, approved, pending, rejected]
            default: all
        - name: category
          in: query
          description: Filter by category ID
          schema:
            type: string
        - name: search
          in: query
          description: |
            Search term to filter results. Matches against title, keywords, and description
            using case-insensitive contains matching. Note: Search is applied after bounded
            fetch, so results are limited to the most recent 200 items per content type.
          schema:
            type: string
        - name: sort
          in: query
          description: Sort order for results
          schema:
            type: string
            enum: [newest, oldest, custom]
            default: newest
        - name: page
          in: query
          description: Page number (0-indexed)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size (max 100)
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Content library page with pagination metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentLibraryResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Authentication JWT token
  schemas:
    HomeCategoryDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        localizedNames:
          type: object
          additionalProperties:
            type: string
          nullable: true
        displayOrder:
          type: integer
          nullable: true
        icon:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContentItemDto'
        totalContentCount:
          type: integer
...
    ContentSortDto:
      type: object
      properties:
        contentId:
          type: string
        contentType:
          type: string
          enum: [channel, playlist, video]
        title:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        position:
          type: integer
        youtubeId:
          type: string
...
    Channel:
      type: object
      required: [id, youtubeId, name]
      properties:
        id:
          type: string
        youtubeId:
          type: string
          description: YouTube channel ID
        name:
          type: string
          description: Channel name/title
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        subscribers:
          type: integer
          format: int64
          nullable: true
          description: Subscriber count
        videoCount:
          type: integer
          nullable: true
          description: Number of videos in this channel
        categoryIds:
          type: array
          items:
            type: string
          description: List of category IDs this channel belongs to
        keywords:
          type: array
          items:
            type: string
          nullable: true
          description: Search keywords/tags for this channel. Auto-populated from YouTube tags during import if not provided.
        status:
          type: string
          enum: [APPROVED, PENDING, REJECTED]
          default: PENDING
          description: Approval status of the channel
        excludedItems:
          $ref: '#/components/schemas/ExcludedItems'
          nullable: true
          description: Items to exclude from this channel
        submittedBy:
          type: string
          nullable: true
          description: UID of user who submitted this channel
        approvedBy:
          type: string
          nullable: true
          description: UID of user who approved this channel
        approvalMetadata:
          $ref: '#/components/schemas/ApprovalMetadata'
          nullable: true
          description: Detailed approval/rejection metadata
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        displayOrder:
          type: integer
          nullable: true
          description: Integer sort order. Lower values appear first. Null until explicitly set.
    Playlist:
      type: object
      required: [id, youtubeId, title]
      properties:
        id:
          type: string
        youtubeId:
          type: string
          description: YouTube playlist ID
        title:
          type: string
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        itemCount:
          type: integer
          nullable: true
          description: Number of videos in this playlist
        categoryIds:
          type: array
          items:
            type: string
          description: List of category IDs this playlist belongs to
        keywords:
          type: array
          items:
            type: string
          nullable: true
          description: Search keywords/tags for this playlist.
        channelId:
          type: string
          nullable: true
          description: ID of the channel that owns this playlist
        status:
          type: string
          enum: [APPROVED, PENDING, REJECTED]
          default: PENDING
          description: Approval status of the playlist
        excludedVideoIds:
          type: array
          items:
            type: string
          nullable: true
          description: YouTube video IDs to exclude from this playlist
        submittedBy:
          type: string
          nullable: true
          description: UID of user who submitted this playlist
        approvedBy:
          type: string
          nullable: true
          description: UID of user who approved this playlist
        approvalMetadata:
          $ref: '#/components/schemas/ApprovalMetadata'
          nullable: true
          description: Detailed approval/rejection metadata
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        displayOrder:
          type: integer
          nullable: true
          description: Integer sort order. Lower values appear first. Null until explicitly set.
    Video:
      type: object
      required: [id, youtubeId, title]
      properties:
        id:
          type: string
        youtubeId:
          type: string
          description: YouTube video ID
        title:
          type: string
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
          description: Video duration in seconds
        uploadedAt:
          type: string
          format: date-time
          nullable: true
          description: When the video was uploaded to YouTube
        viewCount:
          type: integer
          format: int64
          nullable: true
        channelId:
          type: string
          nullable: true
          description: YouTube channel ID that owns this video
        channelTitle:
          type: string
          nullable: true
          description: Channel name (cached from YouTube)
        categoryIds:
          type: array
          items:
            type: string
          description: List of category IDs this video belongs to
        keywords:
          type: array
          items:
            type: string
          nullable: true
          description: Search keywords/tags for this video. Auto-populated from YouTube tags during import if not provided.
        status:
          type: string
          enum: [APPROVED, PENDING, REJECTED, UNAVAILABLE]
          default: PENDING
          description: Approval/availability status of the video
        validationStatus:
          type: string
          enum: [VALID, UNAVAILABLE, ERROR]
          nullable: true
          description: Whether the video still exists on YouTube (VALID=accessible, UNAVAILABLE=deleted/private, ERROR=validation failed, null=not yet validated)
        lastValidatedAt:
          type: string
          format: date-time
          nullable: true
          description: Last time this video was validated against YouTube API
        submittedBy:
          type: string
          nullable: true
          description: UID of user who submitted this video
        approvedBy:
          type: string
          nullable: true
          description: UID of user who approved this video
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        displayOrder:
          type: integer
          nullable: true
          description: Integer sort order. Lower values appear first. Null until explicitly set.
    ExcludedItems:
      type: object
      description: Items to exclude from a channel (videos, playlists, etc.)
      properties:
        videos:
          type: array
          items:
            type: string
          description: YouTube video IDs to exclude
          default: []
        playlists:
          type: array
          items:
            type: string
          description: YouTube playlist IDs to exclude
          default: []
        liveStreams:
          type: array
          items:
            type: string
          description: YouTube livestream IDs to exclude
          default: []
        shorts:
          type: array
          items:
            type: string
          description: YouTube Shorts IDs to exclude
          default: []
        posts:
          type: array
          items:
            type: string
          description: YouTube community post IDs to exclude
          default: []

    # YouTube Search & Preview DTOs
...
    ContentLibraryResponse:
      type: object
      description: |
        Paginated response for content library. The `truncated` field indicates whether
        results may be incomplete due to bounded query limits (max 200 per content type).
        UI should display a warning to users when `truncated` is true.
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ContentItem'
        totalItems:
          type: integer
          description: Total count of items matching filters within the bounded query window
        currentPage:
          type: integer
          description: Current page number (0-indexed)
        pageSize:
          type: integer
          description: Number of items per page
        totalPages:
          type: integer
          description: Total number of pages within the bounded query window
        truncated:
          type: boolean
          description: |
            Indicates whether results may be incomplete due to bounded query limits.
            When true, the server capped results to prevent Firestore quota exhaustion,
            and there may be more items matching the filters than shown. UI should
            display a message like "Showing most recent results. Use filters or search
            to find specific items." when this is true.
    ContentItem:
      type: object
      description: Unified content item (channel/playlist/video) for content library
      properties:
        type:
          type: string
          enum: [channel, playlist, video]
        id:
          type: string
        youtubeId:
          type: string
          description: YouTube ID for video preview/player functionality
        title:
          type: string
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [APPROVED, PENDING, REJECTED, UNAVAILABLE]
        categoryIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
          nullable: true
        count:
          type: integer
          format: int64
          nullable: true
          description: Subscriber/video/view count depending on content type
        displayOrder:
          type: integer
          nullable: true
          description: Integer sort order. Lower values appear first. Null until explicitly set.
        keywords:
          type: array
          items:
            type: string
          nullable: true
          description: Search keywords/tags for improved search accuracy. Auto-populated from YouTube tags during import if not provided.
...
    CursorPageDto:
      type: object
      description: Generic cursor-based pagination wrapper
      properties:
        data:
          type: array
          items:
            type: object
          description: Array of items (type varies by endpoint)
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    PageInfo:
      type: object
      description: Pagination metadata for cursor-based pagination
      required:
        - hasNext
      properties:
        nextCursor:
          type: string
          nullable: true
          description: Cursor for fetching next page (null if last page)
        hasNext:
          type: boolean
          description: Whether more pages are available
        totalCount:
          type: integer
          nullable: true
          description: Total count of items matching current filters (optional, returned by some endpoints)
        truncated:
          type: boolean
          nullable: true
          description: |
            Indicates whether results were truncated due to safety limits.
            When true, the totalCount and results may be incomplete.
            Used by workspace exclusions aggregation to signal when hard limits
            (e.g., 500 channels, 1000 playlists) have been hit.
...
    ApprovalMetadata:
      type: object
      description: Metadata about approval/rejection actions (embedded in Channel/Playlist documents)
      properties:
        reviewedBy:
          type: string
          nullable: true
          description: Firebase UID of the reviewer
        reviewerDisplayName:
          type: string
          nullable: true
          description: Display name of the reviewer (cached)
        reviewedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the review happened
        reviewNotes:
          type: string
          nullable: true
          description: Review notes from the moderator
        rejectionReason:
          type: string
          nullable: true
          enum: [NOT_ISLAMIC, LOW_QUALITY, DUPLICATE, OTHER]
          description: Rejection reason (for rejected items)
...
    ContentItemDto:
      type: object
      description: Unified DTO for content items (channels, playlists, videos) in public API. Fields are nullable to accommodate different content types.
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [CHANNEL, PLAYLIST, VIDEO]
          description: Content type discriminator
        # Common fields
        title:
          type: string
          nullable: true
          description: Title for playlists and videos
        name:
          type: string
          nullable: true
          description: Name for channels
        category:
          type: string
          nullable: true
          description: Category name
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        # Video-specific fields
        durationSeconds:
          type: integer
          nullable: true
          description: Video duration in seconds (videos only)
        uploadedDaysAgo:
          type: integer
          nullable: true
          description: Days since upload (videos only)
        viewCount:
          type: integer
          format: int64
          nullable: true
          description: View count (videos only)
        # Channel-specific fields
        subscribers:
          type: integer
          format: int64
          nullable: true
          description: Subscriber count (channels only)
        videoCount:
          type: integer
          nullable: true
          description: Number of videos (channels only)
        # Playlist-specific fields
        itemCount:
          type: integer
          nullable: true
          description: Number of items (playlists only)
...
    ErrorResponse:
      type: object
      description: Generic error response for 4xx and 5xx errors
      properties:
        error:
          type: string
          description: Error type or code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          nullable: true
          description: Additional error context (optional)
...